+++ streaming/stream.py
@@ -402,7 +402,7 @@ class Stream:
             pass
 
         self.shout.open()
-        self.current_song = self.current_playlist.get_current_song()
+        self.current_song = self.current_playlist.get_first_song()
         print("DID SET CURRENT SONG")
         self._should_announce_next_song()
         self._prepare_next_announcement()